+++ examples/with-next-custom-button/pages/index.tsx
@@ -1,4 +1,4 @@
-import { ConnectButton } from '@rainbow-me/rainbowkit';
+import { ConnectButton } from '@m7eio/rainbowkit';
 import type { NextPage } from 'next';
 import Image from 'next/image';