+++ examples/with-vite/src/main.tsx
@@ -1,7 +1,7 @@
 import './polyfills';
 import './global.css';
-import '@rainbow-me/rainbowkit/styles.css';
-import { getDefaultWallets, RainbowKitProvider } from '@rainbow-me/rainbowkit';
+import '@m7eio/rainbowkit/styles.css';
+import { getDefaultWallets, RainbowKitProvider } from '@m7eio/rainbowkit';
 import { chain, configureChains, createClient, WagmiConfig } from 'wagmi';
 import { alchemyProvider } from 'wagmi/providers/alchemy';
 import { publicProvider } from 'wagmi/providers/public';