+++ examples/with-next-siwe-iron-session/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Updated dependencies [8060ccd]
 - Updated dependencies [4dfe834]
 - Updated dependencies [8060ccd]
-  - @rainbow-me/rainbowkit@0.5.1
+  - @m7eio/rainbowkit@0.5.1
 
 ## 0.0.1
 
@@ -15,4 +15,4 @@
 
 - Updated dependencies [737a1d6]
 - Updated dependencies [488c5a1]
-  - @rainbow-me/rainbowkit@0.5.0
+  - @m7eio/rainbowkit@0.5.0