+++ packages/rainbowkit-siwe-next-auth/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @rainbow-me/rainbowkit-siwe-next-auth
+# @m7eio/rainbowkit-siwe-next-auth
 
 ## 0.1.1
 
@@ -7,7 +7,7 @@
 - Updated dependencies [8060ccd]
 - Updated dependencies [4dfe834]
 - Updated dependencies [8060ccd]
-  - @rainbow-me/rainbowkit@0.5.1
+  - @m7eio/rainbowkit@0.5.1
 
 ## 0.1.0
 
@@ -19,4 +19,4 @@
 
 - Updated dependencies [737a1d6]
 - Updated dependencies [488c5a1]
-  - @rainbow-me/rainbowkit@0.5.0
+  - @m7eio/rainbowkit@0.5.0